pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "boletus"
version = "0.1.0"
description = "Your first AI company — Open-source multi-agent framework that runs your business autonomously"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Majny", email = "boletus@example.com" },
]
keywords = ["ai", "agents", "multi-agent", "autonomous", "slack", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "pypdf>=4.0",
    "python-docx>=1.1.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[project.scripts]
boletus = "boletus.cli:main"

[project.urls]
Homepage = "https://github.com/Majny/boletus"
Repository = "https://github.com/Majny/boletus"
Issues = "https://github.com/Majny/boletus/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120
